Refactor for server/microcontroller/desktop
Further refactoring so there's a Server part which serves the most
heavy-weight api calls in one easy json result. This can be used in
the lite version, so less memory is needed when running this on a
microcontroller. For desktop usage the desktop version uses some of
the other libraries and adds some extra things like printing to screen.